Bugfixes during first startup

+++ RainbowMiner.ps1
@@ -133,7 +133,9 @@ $MSIAcurrentprofile = -1
 $RunSetup = $false
 
 #Cleanup the log
-Get-ChildItem -Path ".\Logs" -Filter "*" | Where-Object {$_.LastWriteTime -lt (Get-Date).AddDays(-5)} | Remove-Item -ErrorAction Ignore
+if (Test-Path ".\Logs"){
+    Get-ChildItem -Path ".\Logs" -Filter "*" | Where-Object {$_.LastWriteTime -lt (Get-Date).AddDays(-5)} | Remove-Item -ErrorAction Ignore
+}
 
 #Start the log
 Start-Transcript ".\Logs\RainbowMiner_$(Get-Date -Format "yyyy-MM-dd_HH-mm-ss").txt"
